backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.10
openai==1.6.1
python-dotenv==1.0.0
schedule==1.2.1
pandas==2.2.0
openpyxl==3.1.2
yfinance==0.2.33
google-generativeai
